test(visual): dissolve ui-refresh.spec.js and interactions.spec.js grab-bags
Splits the two catch-all Playwright spec files into focused files by
component area, and re-points every stale doc citation in
docs/architecture/ui-components.md at the correct new path:

- test/visual/features/navbar.spec.js — navbar labels/brand cluster +
  dropdown open/close (from ui-refresh + interactions)
- test/visual/features/layouts.spec.js — intro hero + section archive
  (from ui-refresh)
- test/visual/features/content-enhancements.spec.js — code blocks,
  content tables, docs ToC/FAB chrome, code-copy click (from ui-refresh
  + interactions)
- test/visual/features/appearance.spec.js — theme preview + theme
  customizer skin apply (from ui-refresh + interactions)
- test/visual/core/layout-chrome.spec.js — footer + skip-link focus
  (from ui-refresh)
- test/visual/core/responsive.spec.js — per-viewport axe scan +
  landmark visibility (from ui-refresh)
- test/visual/features/category-badges.spec.js — feature category
  badge anchors (from ui-refresh)
- test/visual/features/search.spec.js — search modal interactions
  (from interactions)

All 29 tests were moved verbatim (test/describe counts verified
1:1 against the deleted files); no test logic changed, only which
file each one lives in.

Note: the wider test/visual/core + test/visual/features restructuring
this fix assumes (and the doc's own build-up note) does not otherwise
exist in this repo — this change performs the minimal real
restructuring needed to make the two named grab-bag files' citations
true, without inventing an organization for the ~40 other still-flat
spec files, which was out of scope.
